# Sistema_vuelos.v1/vuelos-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Vuelos API
  description: |
    API REST para el módulo de vuelos de la plataforma Booking académica.

    **Versionamiento**: todas las rutas usan el prefijo `/api/v1/`.

    **Autenticación**: Bearer JWT en el header `Authorization`.
    Obtener el token con `POST /api/v1/auth/login`.

    **Validación**: todos los endpoints mutantes (POST/PUT/PATCH) validan el cuerpo con
    Zod antes de ejecutar lógica de negocio. Errores de validación retornan HTTP 400.

    **Integridad de asientos**: la creación de reservas y la cancelación se ejecutan en
    transacciones Prisma atómicas — se garantiza que `availableSeats` nunca queda
    inconsistente aunque haya concurrencia.

    **Integracion con Booking Central**: el sistema expone contratos gRPC en los archivos
    `.proto` de `src/grpc/proto/`. OpenAPI documenta solamente el contrato REST/JSON;
    los metodos gRPC no se declaran aqui para no mezclar protocolos.
  version: 1.1.0
  contact:
    name: Equipo Vuelos
servers:
  - url: https://integracion-sistemas2026.onrender.com/api/v1
    description: Producción (Render)
  - url: http://localhost:3000/api/v1
    description: Desarrollo local

# ============================================================
#   SEGURIDAD GLOBAL
# ============================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en POST /auth/login
  schemas:
...
    SuccessResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:    { type: object }
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrar nuevo usuario
      description: Crea una cuenta con rol CUSTOMER y devuelve token JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, firstLastName, mainAddress]
              properties:
                email:          { type: string, format: email }
                password:       { type: string, minLength: 6 }
                firstName:      { type: string }
                firstLastName:  { type: string }
                secondName:     { type: string }
                secondLastName: { type: string }
                mainAddress:    { type: string }
                cityId:         { type: string, format: uuid }
                phone:          { type: string }
                birthDate:      { type: string, format: date-time }
      responses:
        '201':
          description: Usuario registrado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:  { $ref: '#/components/schemas/User' }
                          token: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:  { $ref: '#/components/schemas/User' }
                          token: { type: string }
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /flights/search:
    get:
      tags: [Flights]
      summary: Buscar vuelos disponibles
      description: |
        Retorna vuelos con disponibilidad para la ruta, fecha y número de pasajeros.
        Solo devuelve vuelos con `availableSeats >= passengers` en la clase solicitada.
      parameters:
        - { name: origin, in: query, required: true, schema: { type: string, example: UIO }, description: Código IATA de 3 letras }
        - { name: destination, in: query, required: true, schema: { type: string, example: BOG } }
        - { name: date, in: query, required: true, schema: { type: string, format: date, example: '2025-06-15' } }
        - { name: passengers, in: query, schema: { type: integer, minimum: 1, maximum: 9, default: 1 } }
        - { name: class, in: query, schema: { type: string, enum: [ECONOMY, PREMIUM_ECONOMY, BUSINESS, FIRST] } }
      responses:
        '200':
          description: Vuelos disponibles
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Flight' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /flights:
    get:
      tags: [Flights]
      summary: Listar todos los vuelos (público)
      responses:
        '200':
          description: Lista de vuelos
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Flight' }
    post:
      tags: [Flights]
      summary: Crear vuelo (admin)
      security: [{ bearerAuth: [] }]
      description: |
        Valida que `originAirportIata ≠ destinationAirportIata`.
        Se pueden incluir segmentos y clases de vuelo en el mismo request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [originAirportIata, destinationAirportIata, departureDate]
              properties:
                originAirportIata:
                  type: string
                  minLength: 3
                  maxLength: 3
                  example: UIO
                destinationAirportIata:
                  type: string
                  minLength: 3
                  maxLength: 3
                  example: BOG
                  description: Debe ser diferente al origen
                departureDate:
                  type: string
                  format: date
                  example: '2025-06-15'
                status:
                  type: string
                  enum: [SCHEDULED, DELAYED, CANCELLED, COMPLETED]
                  default: SCHEDULED
                segments:
                  type: array
                  items:
                    type: object
                    required: [segmentNumber, originAirportId, destinationAirportId, departureDateTime, arrivalDateTime, airlineId, estimatedDuration]
                    properties:
                      segmentNumber:        { type: string }
                      originAirportId:      { type: string, format: uuid }
                      destinationAirportId: { type: string, format: uuid }
                      departureDateTime:    { type: string, format: date-time, description: Debe ser anterior a arrivalDateTime }
                      arrivalDateTime:      { type: string, format: date-time }
                      airlineId:            { type: string, format: uuid }
                      aircraftId:           { type: string, format: uuid }
                      estimatedDuration:    { type: integer, minimum: 1 }
                flightClasses:
                  type: array
                  items:
                    type: object
                    required: [cabinClass, availableSeats, basePrice]
                    properties:
                      cabinClass:     { type: string, enum: [ECONOMY, PREMIUM_ECONOMY, BUSINESS, FIRST] }
                      availableSeats: { type: integer, minimum: 1 }
                      basePrice:      { type: number, minimum: 0, exclusiveMinimum: true }
      responses:
        '201':
          description: Vuelo creado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Flight' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /flight-classes:
    get:
      tags: [FlightClasses]
      summary: Listar clases de vuelo
      responses:
        '200':
          description: Lista de clases
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/FlightClass' }
    post:
      tags: [FlightClasses]
      summary: Crear clase de vuelo (admin)
      security: [{ bearerAuth: [] }]
      description: |
        Única por combinación `(flightId, cabinClass)` — retorna 409 si ya existe.
        `availableSeats` mínimo 0; `basePrice` debe ser mayor a 0.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [flightId, cabinClass, availableSeats, basePrice]
              properties:
                flightId:       { type: string, format: uuid }
                cabinClass:     { type: string, enum: [ECONOMY, PREMIUM_ECONOMY, BUSINESS, FIRST] }
                availableSeats: { type: integer, minimum: 0 }
                basePrice:      { type: number, minimum: 0, exclusiveMinimum: true }
      responses:
        '201':
          description: Clase creada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/FlightClass' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /reservations:
    post:
      tags: [Reservations]
      summary: Crear reserva
      description: |
        Transacción atómica que:
        1. Verifica disponibilidad (`availableSeats >= pasajeros`)
        2. Aplica descuento si se envía `promotionCode` (valida existencia, estado activo y límite de usos)
        3. Crea la reserva y sus pasajeros en un solo commit
        4. Decrementa `availableSeats` e incrementa `currentUsages` de la promo

        Si dos peticiones concurrentes compiten por el último asiento, solo una confirma.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [flightClassId, passengers]
              properties:
                flightClassId:
                  type: string
                  format: uuid
                passengers:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [firstName, lastName, documentNumber]
                    properties:
                      firstName:      { type: string }
                      lastName:       { type: string }
                      documentNumber: { type: string }
                      seatNumber:     { type: string, nullable: true }
                promotionCode:
                  type: string
                  example: VERANO25
      responses:
        '201':
          description: Reserva confirmada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Reservation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422':
          description: Sin disponibilidad
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    get:
      tags: [Reservations]
      summary: Listar todas las reservas (admin)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Lista de reservas
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Reservation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /payments:
    post:
      tags: [Payments]
      summary: Registrar pago
      description: |
        Crea el registro de pago y dispara `autoGenerateInvoice()` de forma asíncrona.
        La factura se genera automáticamente con IVA 15% basado en el monto del pago.
        El número de factura usa `crypto.randomBytes` para evitar colisiones.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reservationId, amount, provider, transactionId]
              properties:
                reservationId: { type: string, format: uuid }
                amount:        { type: number, minimum: 0, exclusiveMinimum: true }
                provider:      { type: string, example: VISA }
                transactionId: { type: string, example: TXN-20250615-001 }
                status:        { type: string, enum: [PENDING, COMPLETED, FAILED, REFUNDED] }
      responses:
        '201':
          description: Pago registrado. La factura se genera de forma asíncrona.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Payment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    get:
      tags: [Payments]
      summary: Listar pagos (admin)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Lista de pagos
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Payment' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /invoices:
    get:
      tags: [Invoices]
      summary: Listar facturas (admin)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Lista de facturas
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Invoice' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /invoice-items:
    post:
      tags: [Invoices]
      summary: Crear ítem de factura (admin)
      description: |
        Todos los campos son validados por Zod antes de persistir.
        `totalPrice` debe ser coherente con `quantity * unitPrice`.
        Después de crear/modificar ítems, el frontend recalcula automáticamente
        subtotal, taxAmount y total de la factura padre.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invoiceId, description, quantity, unitPrice, totalPrice]
              properties:
                invoiceId:   { type: string, format: uuid }
                description: { type: string, minLength: 1 }
                quantity:    { type: integer, minimum: 1 }
                unitPrice:   { type: number, minimum: 0 }
                totalPrice:  { type: number, minimum: 0 }
      responses:
        '201':
          description: Ítem creado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/InvoiceItem' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /admin/promotions:
    get:
      tags: [Admin]
      summary: Listar promociones (admin CRUD)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Lista, content: { application/json: { schema: { $ref: '#/components/schemas/SuccessResponse' } } } }
  /admin/invoice-items:
    get:
      tags: [Admin]
      summary: Listar ítems de factura
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Lista de ítems, content: { application/json: { schema: { $ref: '#/components/schemas/SuccessResponse' } } } }
...